requirements.txt
pandas==2.0.2
tensorflow==2.12.0
matplotlib==3.7.1
scipy==1.10.1
scikit-learn==1.2.2
tensorflow_probability==0.20.1
pyelftools
